+++ app/src/main/kotlin/com/stevenfrew/beatprompter/storage/StorageListener.kt
@@ -4,6 +4,5 @@ package com.stevenfrew.beatprompter.storage
  * Listener for general storage features.
  */
 interface StorageListener {
-	fun onAuthenticationRequired()
 	fun shouldCancel(): Boolean
 }